+++ CHANGELOG.md
@@ -33,6 +33,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
+## Unreleased
+
+ - [#189](https://github.com/babylonlabs-io/btc-staker/pull/189) Use signing context
+ string in staker pop
+
 ## v0.16.0-rc.0
 
 ### Improvements